So, it turns out Node's `readline` implementation is fairly slow (at least, on my machine). The largest change in this PR can basically be summed up as:

1. Refactor the fast space-finding code out to a function that iterates over substrings terminated by a given token, `forEachTerminated`, since the previous optimized version proved that this is faster than calling `String#split` and then iterating over the returned array.
2. Read the `data` events directly from `process.stdin` rather than through the `readline` abstraction. Use `forEachTerminated` to pull out newline-terminated substrings — since `data` will be arbitrary chunks of utf-8 from stdin, and not specifically *lines* of stdin — and hand those to `handleLine`.

I also removed the class abstraction in favor of ordinary dicts with `Object.entries` for iteration, since the class didn't seem to do anything in terms of performance compared to using a dictionary of words to numbers. `Object.entries` gives any cache coherency wins and hashing-function skipping wins that you'd get from the class abstraction, and in practice it takes less code to do this with a dict of words to numbers rather than setting up the special class. The simple version has actually already switched to `Object.entries` (and got a small perf boost out of it), but it looks like the optimized version was written before that switch. There might be a minor speed boost with `Object.entries` here too due to creating less garbage for the GC to examine, not rehashing strings during `sort`, and avoiding chasing pointers to random memory locations when doing key lookups in the dict, but I didn't see a statistically significant difference on my machine when measuring that change with `hyperfine` — it just was simpler and required fewer lines of code. The only perf win I could consistently measure was from avoiding `readline`.

Overall this PR is about 15-20% faster on my machine due to avoiding `readline`, and the end result takes fewer lines of code due to removing the class abstraction!