@benhoyt 2021-03-17
Thanks! Sounds reasonable to me at a glance. I'll test later today and update the results.