@Daninet 2021-03-17
Nice optimization @reissbaker. On my computer (i7-7700k, Windows 10 x64, Node.js v14.15.3) it is about 10% faster compared to the previous code.

On my computer the "RefNum" and the Object.keys() are measurably faster, I'm consistently getting about 5% improvement for that. I tried keeping the input data as a buffer and it gave me additional 5%. But I guess these microoptimizations are very dependent on the platform / CPU / OS / Node.js version.

Here is my code which gives me an additional 10% over your code:
```js
"use strict";

const dict = {};

// Calls itemHandler on each substring terminated by the token.
// Returns any leftover characters in the string
const forEachTerminated = (str, token, itemHandler) => {
  let lastIndex = 0;

  while (true) {
    const index = str.indexOf(token, lastIndex);

    if (index === -1) break;
    itemHandler(str.slice(lastIndex, index));

    lastIndex = index + 1;
  }

  return str.slice(lastIndex);
};

function RefNum(num) {
  this.num = num;
}

const wordHandler = (word) => {
  if (word.length === 0) return;
  const item = dict[word];
  if (item !== undefined) {
    item.num++;
  } else {
    dict[word] = new RefNum(1);
  }
};

const lineHandler = (line) => {
  if (line.length === 0) return;
  wordHandler(forEachTerminated(line, " ", wordHandler).toString());
};

const endHandler = () => {
  const keys = Object.keys(dict);
  keys.sort((a, b) => dict[b].num - dict[a].num);

  process.stdout.write(keys.map((key) => `${key} ${dict[key].num}\n`).join(""));
};

let buffer = Buffer.allocUnsafe(0);
process.stdin.resume();

process.stdin.on("data", (data) => {
  const str = Buffer.concat([buffer, data]).toString().toLowerCase();
  buffer = Buffer.from(forEachTerminated(str, "\n", lineHandler));
});

process.stdin.on("end", () => {
  lineHandler(buffer.toString());
  endHandler();
});
```