@reissbaker 2021-03-17
Nice @Daninet, I'll definitely try adding the Buffer changes! I'll try re-benchmarking the class version as well. It does seem like it could be a tossup depending on Node versions and OS versions (I am in the particularly weird case of running Node on Ubuntu inside WSL2 on Windows, so there could definitely be something pretty weird about my setup); the class approach trades off writes to the dictionary with more allocations and potentially GC pressure. I would naively also expect `Object.entries` to have better cache coherency than `Object.keys` + dictionary lookups for sorting and transforming the output, but JS "arrays" are magical black boxes so that could be completely wrong.

One question: how are you measuring the changes? I use `hyperfine` with large iteration counts, and even then there's typically enough deltas between runs that it would struggle I think to conclusively decide between implementations that are 5% different.