@reissbaker 2021-03-17
Odd findings for my machine, at least:

* There didn't seem to be a conclusive difference between the class version and the classless version. The class version was ever so slightly faster on average across 30 runs, but `hyperfine` reported `±  0.164 s` and the difference was smaller than that. Since it sounds like it's faster for you, I'll put up a PR tomorrow with the change (or feel free to do it if you get to it before I do!) and we'll see if it's faster on the ultimate test machine, which is @benhoyt's since that's what results in the numbers getting published to the blog post heh.
* The Buffer changes were actually about 10% worse on my machine! Intuitively that makes at least some sense to me: `Buffer.from` is presumably doing some extra copying, and I would expect setting `setEncoding('utf-8')` once would be faster than calling `.toString()` from JavaScript repeatedly — but I'm just treating V8 as a black box, and I don't know what the internals truly are. It seems like this is probably something that varies across platforms if it's significantly faster for you, since it did seem measurably slower for me.